Ticket #47378 - fix recent compiler warnings
https://fedorahosted.org/389/ticket/47378
Reviewed by: nhosoi (Thanks!)
Branch: master
Fix Description: Fix the various compiler warnings as reported by
http://jenkins.cloud.fedoraproject.org/job/389-ds-base/ws/build-warns.1370449480.txt
In addition, fixed a bug in the slapi_re_subs_ext - if there was an
ampersand that did not need to be replaced, that is, if the ampersand
was part of an LDAP search filter, it was replaced with a '('.  Also, an
ampersand preceeded by an open parenthesis and followed by a non-open
parenthesis is supposed to be replaced in the string, but this was not
working.
Platforms tested: RHEL6 x86_64
Flag Day: no
Doc impact: no
(cherry picked from commit 0dc230b007dc3ad292961f378421b03736b68f9e)

+++ ldap/servers/slapd/back-ldbm/monitor.c
@@ -238,13 +238,13 @@ int ldbm_back_monitor_instance_search(Slapi_PBlock *pb, Slapi_Entry *e,
 #endif
         MSETF("dbFilename-%d", i);
         
-        sprintf(buf, "%u", mpfstat[i]->st_cache_hit);
+        sprintf(buf, "%lu", (unsigned long)mpfstat[i]->st_cache_hit);
         MSETF("dbFileCacheHit-%d", i);
-        sprintf(buf, "%u", mpfstat[i]->st_cache_miss);
+        sprintf(buf, "%lu", (unsigned long)mpfstat[i]->st_cache_miss);
         MSETF("dbFileCacheMiss-%d", i);
-        sprintf(buf, "%u", mpfstat[i]->st_page_in);
+        sprintf(buf, "%lu", (unsigned long)mpfstat[i]->st_page_in);
         MSETF("dbFilePageIn-%d", i);
-        sprintf(buf, "%u", mpfstat[i]->st_page_out);
+        sprintf(buf, "%lu", (unsigned long)mpfstat[i]->st_page_out);
         MSETF("dbFilePageOut-%d", i);
 
 	slapi_ch_free_string(&absolute_pathname);
@@ -291,7 +291,7 @@ int ldbm_back_monitor_search(Slapi_PBlock *pb, Slapi_Entry *e,
     }
 
     /* cache hits*/
-    sprintf(buf, "%u", mpstat->st_cache_hit);
+    sprintf(buf, "%lu", (unsigned long)mpstat->st_cache_hit);
     MSET("dbCacheHits");
 
     /* cache tries*/
@@ -303,13 +303,13 @@ int ldbm_back_monitor_search(Slapi_PBlock *pb, Slapi_Entry *e,
     sprintf(buf, "%lu", (unsigned long)(100.0 * (double)mpstat->st_cache_hit / (double)(cache_tries > 0 ? cache_tries : 1) ));
     MSET("dbCacheHitRatio");
 
-    sprintf(buf, "%u", mpstat->st_page_in);
+    sprintf(buf, "%lu", (unsigned long)mpstat->st_page_in);
     MSET("dbCachePageIn");
-    sprintf(buf, "%u", mpstat->st_page_out);
+    sprintf(buf, "%lu", (unsigned long)mpstat->st_page_out);
     MSET("dbCachePageOut");
-    sprintf(buf, "%u", mpstat->st_ro_evict);
+    sprintf(buf, "%lu", (unsigned long)mpstat->st_ro_evict);
     MSET("dbCacheROEvict");
-    sprintf(buf, "%u", mpstat->st_rw_evict);
+    sprintf(buf, "%lu", (unsigned long)mpstat->st_rw_evict);
     MSET("dbCacheRWEvict");
 
     slapi_ch_free((void **)&mpstat);
